--- src/main/java/kitchenpos/application/ProductService.java
@@ -1,27 +1,0 @@
-package kitchenpos.application;
-
-import java.util.List;
-import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Product;
-import kitchenpos.application.dto.ProductCreateRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class ProductService {
-    private final ProductDao productDao;
-
-    public ProductService(final ProductDao productDao) {
-        this.productDao = productDao;
-    }
-
-    @Transactional
-    public Product create(final ProductCreateRequest request) {
-        Product product = new Product(request.getName(), request.getPrice());
-        return productDao.save(product);
-    }
-
-    public List<Product> list() {
-        return productDao.findAll();
-    }
-}
