+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,0 +1,27 @@
+package kitchenpos.product.application;
+
+import java.util.List;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductCreateRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class ProductService {
+    private final ProductRepository productRepository;
+
+    public ProductService(final ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Transactional
+    public Product create(final ProductCreateRequest request) {
+        Product product = new Product(request.getName(), request.getPrice());
+        return productRepository.save(product);
+    }
+
+    public List<Product> list() {
+        return productRepository.findAll();
+    }
+}
